% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addCustAnnotation.R
\name{addCustAnnotation}
\alias{addCustAnnotation}
\title{adds a custom annotation to an object
\code{addCustannotation} puts a file to a namespace}
\usage{
addCustAnnotation(fileName, namespace, auth, filePath, annotationName,
  verbose = FALSE)
}
\arguments{
\item{namespace}{namespace to get object list from}

\item{auth}{authentication string}

\item{filePath}{path to annotation file}

\item{annotationName}{name for the custom annotation slot}

\item{verbose}{Print a messages default = FALSE}

\item{filename}{name for object in namespace}
}
\value{
returns character array of object names if sucessful, Otherwise returms "Error" followed by the returned http headers.
}
\description{
adds a custom annotation to an object
\code{addCustannotation} puts a file to a namespace
}
\examples{
\dontrun{addCustannotation(namespace,auth)}
}

