% Generated by roxygen2 (4.0.1): do not edit by hand
\name{authString}
\alias{authString}
\title{Create authorization string for HCP access
\code{authString} Creates encoded authorization string using base64enc and digest packages}
\usage{
authString(username, password)
}
\arguments{
\item{username}{username authorized for namespace}

\item{password}{password for username}
}
\value{
returns authorization text string to be included Http header
}
\description{
Create authorization string for HCP access
\code{authString} Creates encoded authorization string using base64enc and digest packages
}
\examples{
authString(readline("username - "),readline("password - "))
}

