% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createDir.R
\name{createDir}
\alias{createDir}
\title{Creates a directory in a namespace
\code{createDirectory}create a directory in a namespace}
\usage{
createDir(dirName, namespace, auth, verbose = FALSE)
}
\arguments{
\item{dirName}{name for the new directory}

\item{namespace}{namespace to get object list from}

\item{auth}{authorization string created with authString}

\item{verbose}{Print a messages default = FALSE}
}
\value{
r
}
\description{
Creates a directory in a namespace
\code{createDirectory}create a directory in a namespace
}
\examples{
\dontrun{}
}

