% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getCustAnnotation.R
\name{getCustAnnotation}
\alias{getCustAnnotation}
\title{gets a custom annotation of an object from HCP
\code{getCustAnnotation} gets a object (file) a namespace}
\usage{
getCustAnnotation(objectName, annotationName, saveName, namespace, auth,
  verbose = FALSE)
}
\arguments{
\item{objectName}{path to file to be uploaded}

\item{annotationName}{name of stored custome annotation}

\item{saveName}{name for annotation}

\item{namespace}{namespace to get object list from}

\item{auth}{authentication string}

\item{verbose}{Print a messages default = FALSE}
}
\value{
returns "annotation saved in ..." if sucessful, "Error" followed by the returned http headers.
}
\description{
gets a custom annotation of an object from HCP
\code{getCustAnnotation} gets a object (file) a namespace
}
\examples{
\dontrun{getCustAnnotation(objectName,annotationName,saveName,namespace,auth,verbose=True)}
}

