% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getObject.R
\name{getObject}
\alias{getObject}
\title{gets a object (file) from HCP without annotations
\code{authString} gets a object (file) a namespace}
\usage{
getObject(objectName, saveName, namespace, auth, compress = FALSE,
  validate = TRUE, verbose = FALSE)
}
\arguments{
\item{objectName}{path to file to be uploaded}

\item{saveName}{name for object in namespace}

\item{namespace}{namespace to get object list from}

\item{auth}{authentication string}

\item{compress}{use gzip compression for upload default = FALSE}

\item{validate}{validate file checksum}

\item{verbose}{Print a messages default = FALSE}
}
\value{
returns "File saved" if sucessful, "Error" followed by the returned http headers.
}
\description{
gets a object (file) from HCP without annotations
\code{authString} gets a object (file) a namespace
}
\examples{
\dontrun{getOject(objectName,saveName,namespace,auth,compress=TRUE,verbose=True)}
}

