% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HCPtools.R
\docType{package}
\name{HCPtools}
\alias{HCPtools}
\alias{HCPtools-package}
\title{HCPtools  Tools for using the REST APIs on the Hitachi Content Platform}
\description{
HCPtools  Tools for using the REST APIs on the Hitachi Content Platform
}

