% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/listToXML.R
\name{listToXML}
\alias{listToXML}
\title{Create  XML from a list object for a custom HCP annotation
\code{listToXML} Creates encoded authorization string using base64enc and digest packages}
\code{listToXML} Creates encoded authorization string using base64enc and digest packages
\code{listToXML} Creates  XML from a list object for a custom HCP annotation
\code{authString} Creates encoded authorization string using base64enc and digest packages}
\usage{
listToXML(rootname, sublist)
}
\arguments{
\item{node}{name of root node in XML}

\item{list}{object to convert}
}
\value{
XML object
}
\description{
Create  XML from a list object for a custom HCP annotation
\code{listToXML} Creates encoded authorization string using base64enc and digest packages
}
\examples{
\dontrun{listToXML("rootname",list(field1="a",field2="b"))}
\code{listToXML} Creates  XML from a list object for a custom HCP annotation
\code{authString} Creates encoded authorization string using base64enc and digest packages
}

