% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xmlFileQuery.R
\name{xmlFileQuery}
\alias{xmlFileQuery}
\title{queryies HCP with an XML file
\code{xmlFileQuery} gets a object (file) a namespace}
\usage{
xmlFileQuery(xmlFile, namespace, auth, verbose = FALSE)
}
\arguments{
\item{xmlFile}{file path to xml query file}

\item{namespace}{namespace to get object list from}

\item{auth}{authentication string}

\item{verbose}{Print a messages default = FALSE}
}
\value{
returns XML object if sucessful or "Error" followed by the returned http headers.
}
\description{
queryies HCP with an XML file
\code{xmlFileQuery} gets a object (file) a namespace
}
\examples{
\dontrun{xmlFileQuery(xmlFile,saveName,namespace,auth,verbose=True)}
}

