% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deleteDir}
\alias{deleteDir}
\title{Deletes an empty directory in a namespace
\code{deleteDirectory} delete a directory in a namespace}
\usage{
deleteDir(dirName, namespace, auth, verbose = FALSE)
}
\arguments{
\item{dirName}{name of the directory to delete}

\item{auth}{authorization string created with authString}

\item{namespace}{namespace to get object list from}

\item{verbose}{Print a messages default = FALSE}
}
\value{
Upon success "Directory sucessfully deleted", else erroe message
}
\description{
Deletes an empty directory in a namespace
\code{deleteDirectory} delete a directory in a namespace
}
\examples{
\dontrun{deleteDir("testdir3",namespace,auth)}
}

