% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deleteObject.R
\name{deleteObject}
\alias{deleteObject}
\title{Deletes an object in a namespace
\code{deleteObject} delete an object in a namespace}
\usage{
deleteObject(objectName, namespace, auth, verbose = FALSE)
}
\arguments{
\item{namespace}{namespace to get object list from}

\item{auth}{authorization string created with authString}

\item{verbose}{Print a messages default = FALSE}

\item{ObjectName}{name of the directory to delete}
}
\value{
Upon success "Object sucessfully deleted", else erroe message
}
\description{
Deletes an object in a namespace
\code{deleteObject} delete an object in a namespace
}
\examples{
\dontrun{deleteObject("myFile",namespace,auth)}
}

