% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getObjectsList}
\alias{getObjectsList}
\title{get current objects a directory
\code{authString} CReturns list of current object in a directory}
\usage{
getObjectsList(namespace, auth, verbose = FALSE)
}
\arguments{
\item{namespace}{namespace to get object list from}

\item{auth}{authorization string created with authString}

\item{verbose}{Print a messages default = FALSE}
}
\value{
returns charactor array of object names if sucessful,
Otherwise returms "Error" followed by the returned http headers.
}
\description{
get current objects a directory
\code{authString} CReturns list of current object in a directory
}
\examples{
\dontrun{getObjectsList(namespace,auth)}
}

