<<<<<<< HEAD
% Generated by roxygen2 (4.0.2): do not edit by hand
=======
% Generated by roxygen2 (4.0.1): do not edit by hand
>>>>>>> a9d90952f4a04e8ce175224380592c76e4459d01
\name{getObjectsList}
\alias{getObjectsList}
\title{get current objects a directory
\code{authString} CReturns list of current object in a directory}
\usage{
getObjectsList(namespace, auth, verbose = FALSE)
}
\arguments{
<<<<<<< HEAD
\item{namespace}{namespace to get object list from}

\item{auth}{authorization string created with authString}

=======
\item{auth}{authorization string created with authString}

\item{namespace}{namespace to get object list from}

>>>>>>> a9d90952f4a04e8ce175224380592c76e4459d01
\item{verbose}{Print a messages default = FALSE}
}
\value{
returns charactor array of object names if sucessful,
Otherwise returms "Error" followed by the returned http headers.
}
\description{
get current objects a directory
\code{authString} CReturns list of current object in a directory
}
\examples{
\dontrun{getObjectsList(namespace,auth)}
}

