% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/listToS3XML.R
\name{listToS3XML}
\alias{listToS3XML}
\title{Create  XML from a list object for a custom HCP annotation that is comaptable with S3 browers
\code{listToS3XML} Creates encoded authorization string using base64enc and digest packages}
\usage{
listToS3XML(sublist)
}
\arguments{
\item{node}{name of root node in XML}

\item{list}{object to convert}
}
\description{
Create  XML from a list object for a custom HCP annotation that is comaptable with S3 browers
\code{listToS3XML} Creates encoded authorization string using base64enc and digest packages
}
\examples{
\dontrun{listToS3XML(list(field1="a",field2="b"))}
}

