% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/putObject.R
\name{putFile}
\alias{putFile}
\title{puts a file to HCP
\code{putFile} puts a file to a namespace}
\usage{
putFile(filePath, fileName, namespace, auth, compress = FALSE,
  verbose = FALSE)
}
\arguments{
\item{filePath}{path to file to be uploaded}

\item{namespace}{namespace to get object list from}

\item{auth}{authentication string}

\item{compress}{use gzip compression for upload default = FALSE}

\item{verbose}{Print a messages default = FALSE}

\item{filename}{name for object in namespace}
}
\value{
returns charactor array of object names if sucessful, Otherwise returms "Error" followed by the returned http headers.
}
\description{
puts a file to HCP
\code{putFile} puts a file to a namespace
}
\examples{
\dontrun{getObjectsList(namespace,auth)}
}

